Conflicts:
	archive/dexp.py
	archive/local.py

+++ archive/dexp.py
@@ -445,7 +445,11 @@ class ImageTable(Table):
         # Can't have NaNs in int column
         try: hdr['HPIX'] = int(ang2pix(4096,hdr['RA_CENT'],hdr['DEC_CENT']))
         except KeyError: hdr['HPIX'] = -1
-         
+
+        # NAXIS1 is changed due to compression
+        hdr['NAXIS1'] = 2048
+        hdr['NAXIS2'] = 4096
+
         return hdr
 
     def delete_image(self,filename):
@@ -744,7 +748,7 @@ class ProctagTable(Table):
         self.tags = yaml.load(open(self._tags))
 
     def load_proctag(self, tag, query=None, expnum=None):
-        """ Create the proctag and load to db. Use tag lookup query.
+        """ Create the proctag and load to db. Use just 'tag' to lookup query.
 
         Parameters:
         -----------
@@ -761,7 +765,7 @@ class ProctagTable(Table):
         return proctag
 
     def create_proctag(self, tag, query=None, expnum=None):
-        """ Create the proctag data array.
+        """ Create the proctag data array from a tag, query, or expnum
         
         Parameters:
         -----------
@@ -1040,7 +1044,8 @@ def finalize_table(cls):
     tab.create_indexes()
     tab.grant_table()
 
-def load_exposure_table(expnum=None,chunk_size=100,multiproc=False,force=False):
+def load_exposure_table(expnum=None,chunk_size=100,multiproc=False,force=False,
+                        dryrun=False, paths=None):
     """Load the exposure table from the raw file inventory on disk.
 
     Parameters:
@@ -1059,11 +1064,11 @@ def load_exposure_table(expnum=None,chunk_size=100,multiproc=False,force=False):
 
     if force and len(expnum) and expnum[0] is not None:
         logging.info("Removing files for %i exposure(s)..."%len(expnum))
-        tab.delete_by_expnum(expnum)
+        if not dryrun: tab.delete_by_expnum(expnum)
 
     # No exposures specified, get everything from disk
     if not len(expnum) or expnum[0] is None:
-        inv = archive.local.get_inventory()
+        inv = archive.local.get_inventory(paths)
         expnum = np.unique(inv['expnum'])
 
     # Exposures that are not loaded
@@ -1075,9 +1080,11 @@ def load_exposure_table(expnum=None,chunk_size=100,multiproc=False,force=False):
         return
 
     logging.debug("Loading %i exposure(s)..."%len(expnum))
-    return tab.load_chunks(expnum,chunk_size)
+    if not dryrun:
+        return tab.load_chunks(expnum,chunk_size)
 
-def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False):
+def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False,
+                       dryrun=False,paths=None):
     """Load the file archive table from the reduced file inventory on disk.
 
     Parameters:
@@ -1086,6 +1093,7 @@ def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False):
     chunk_size : size of chunk of exposures to load
     multiproc  : number of multiprocessing cores to use
     force      : overwrite of existing file info
+    dryrun     : don't execute anything
 
     Returns:
     --------
@@ -1097,7 +1105,7 @@ def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False):
     # If 'expnum' and 'force', delete existing exposures
     if force and len(expnum) and expnum[0] is not None:
         logging.info("Removing files for %i exposure(s)..."%len(expnum))
-        tab.delete_by_expnum(expnum)
+        if not dryrun: tab.delete_by_expnum(expnum)
 
     # Get missing exposure numbers
     inv = tab.get_missing_expnum()
@@ -1121,11 +1129,11 @@ def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False):
 
         # Get all the files
         files = archive.local.get_reduced_files(expnum=chunk,suffix='_*.fits*',
-                                                multiproc=multiproc)
+                                                multiproc=multiproc,paths=paths)
         psffiles = archive.local.get_psfex_files(expnum=chunk,
-                                                 multiproc=multiproc)
+                                                 multiproc=multiproc,paths=paths)
         zpfiles = archive.local.get_zeropoint_files(expnum=chunk,
-                                                    multiproc=multiproc)
+                                                    multiproc=multiproc,paths=paths)
 
         filepath = np.unique(np.concatenate([files,zpfiles,psffiles]))
      
@@ -1138,11 +1146,12 @@ def load_archive_table(expnum=None,chunk_size=1e3,multiproc=True,force=False):
                where e.expnum = a.expnum and a.band is NULL
             """%dict(archive=tab.tablename,exposure=ExposureTable().tablename)
     logging.debug(query)
-    tab.db.execute(query)
+    if not dryrun: tab.db.execute(query)
 
     return
 
-def load_table(cls, expnum=None, chunk_size=100, multiproc=True,force=False):
+def load_table(cls, expnum=None, chunk_size=100, multiproc=True, force=False,
+               dryrun=False, paths=None):
     """Standalone function to load a table from files.
 
     Parameters:
@@ -1151,6 +1160,8 @@ def load_table(cls, expnum=None, chunk_size=100, multiproc=True,force=False):
     chunk_size : size of chunk to upload
     multiproc  : use multiprocessing?
     force      : delete and reload existing expnum
+    dryrun     : don't execute anything
+    paths      : path(s) to search for expnum
 
     Returns:
     --------
@@ -1162,9 +1173,9 @@ def load_table(cls, expnum=None, chunk_size=100, multiproc=True,force=False):
     # If 'force' and 'expnum' specified, delete files that are already loaded
     if force and len(expnum) and expnum[0] is not None:
         logging.info("Removing files for %i exposure(s)..."%len(expnum))
-        tab.delete_by_expnum(expnum)
+        if not dryrun: tab.delete_by_expnum(expnum)
 
-    # Upload all files that are 'processed'
+    # Query for all files that are 'processed'
     logging.debug("Getting files from database...")
     inv = tab.get_missing_filepaths()
     filepath = inv['filepath']
@@ -1178,9 +1189,12 @@ def load_table(cls, expnum=None, chunk_size=100, multiproc=True,force=False):
         return
 
     logging.debug("Loading %s file(s)..."%len(filepath))
-    return tab.load_chunks(filepath,chunk_size)
 
-def load_zeropoint_table(expnum=None,chunk_size=5e3,multiproc=True,force=False):
+    if not dryrun:
+        return tab.load_chunks(filepath,chunk_size)
+
+def load_zeropoint_table(expnum=None,chunk_size=5e3,multiproc=True,force=False,
+                         dryrun=False, paths=None):
     """Load the zeropoint table.
 
     Parameters:
@@ -1189,14 +1203,18 @@ def load_zeropoint_table(expnum=None,chunk_size=5e3,multiproc=True,force=False):
     chunk_size : size of chunk to upload
     multiproc  : use multiprocessing?
     force      : delete and reload existing expnum
+    dryrun     : don't execute anything
+    paths      : path(s) to search for expnum
 
     Returns:
     --------
     None
     """
-    return load_table(ZeropointTable,expnum,chunk_size,multiproc,force)
+    return load_table(ZeropointTable,expnum,chunk_size,multiproc,force,
+                      dryrun=dryrun,paths=paths)
 
-def load_image_table(expnum=None,chunk_size=5e3,multiproc=True,force=False):
+def load_image_table(expnum=None,chunk_size=5e3,multiproc=True,force=False,
+                     dryrun=False, paths=None):
     """Load the image table.
 
     Parameters:
@@ -1205,12 +1223,15 @@ def load_image_table(expnum=None,chunk_size=5e3,multiproc=True,force=False):
     chunk_size : size of chunk to upload
     multiproc  : use multiprocessing?
     force      : delete and reload existing expnum
+    dryrun     : don't execute anything
+    paths      : path(s) to search for expnum
 
     Returns:
     --------
     None
     """
-    return load_table(ImageTable,expnum,chunk_size,multiproc,force)
+    return load_table(ImageTable,expnum,chunk_size,multiproc,force,
+                      dryrun=dryrun, paths=paths)
 
 def load_object_table(filepath=None,chunk_size=100,force=False):
     """Standalone function to load the object table.
@@ -1243,7 +1264,11 @@ def load_object_table(filepath=None,chunk_size=100,force=False):
 
     logging.debug("Loading objects...")
     return obj.load_catalogs(filepath,chunk_size,force)
-        
+
+def load_proctag_table(tag,query=None,expnum=None):
+    proc = ProctagTable()
+    return proc.load_proctag(tag, query, expnum)
+
 if __name__ == "__main__":
     import argparse
     parser = argparse.ArgumentParser(description=__doc__)
